refactor: split iron proxy backend lifecycle
